+++ src/main/java/eu/printingin3d/smalogger/api/inverterdata/IInverterCommand.java
@@ -1,7 +1,7 @@
 package eu.printingin3d.smalogger.api.inverterdata;
 
 public interface IInverterCommand {
-	long getCommand();
-	long getFirst();
-	long getLast();
+	int getCommand();
+	int getFirst();
+	int getLast();
 }
